fixed up buttons and styling

+++ app/main/main_routes.py
@@ -1,4 +1,4 @@
-from flask import render_template, flash, redirect, url_for, request, session
+from flask import render_template, flash, redirect, url_for, request, abort, session
 from .main_forms import LoginForm, RegistrationForm
 from flask_login import current_user, login_user, logout_user, login_required
 from urllib.parse import urlparse
@@ -74,12 +74,14 @@ def register():
       return redirect(url_for('.login'))
     return render_template('register.html', form=form)
 
-@bp.route('/new', methods=['GET'])
+@bp.route('/new')
+@login_required
 def new_task():
   task = Task()
   return render_template('new_task.html', task=task, action_url = url_for('main.add_task', id=task.id))
 
 @bp.route('/add>', methods=['POST'])
+@login_required
 def add_task():
     task = Task()
     task.name = request.form['name']
@@ -91,14 +93,22 @@ def add_task():
     db.session.commit() 
     return redirect(url_for('.home'))
 
-@bp.route('/edit/<id>', methods=['GET', 'POST'])
+@bp.route('/edit/<id>')
+@login_required
 def edit_task(id):
     task = Task.query.get(id)
+    if not task or task.user_id != current_user.id:       
+      abort(403)
+       
     return render_template('edit_task.html', task=task, action_url = url_for('main.update_task', id=task.id))
     
 @bp.route('/update/<id>', methods=['POST'])
+@login_required
 def update_task(id):
     task = Task.query.get(id)
+    if not task or task.user_id != current_user.id:       
+      abort(403)
+
     task.name = request.form['name']
     task.description = request.form['description']
     task.category = request.form['category']
@@ -108,16 +118,24 @@ def update_task(id):
     db.session.commit() 
     return redirect(url_for('.home'))
 
-@bp.route('/completed/<id>', methods=['GET', 'POST'])
+@bp.route('/complete/<id>', methods=['POST'])
+@login_required
 def complete_task(id):
-    task = Task.query.get(id)
+    task = Task.query.get(id)    
+    if not task or task.user_id != current_user.id:       
+      abort(403)
+
     task.done()
     db.session.commit()
     return redirect(url_for('.home'))
 
-@bp.route('/delete/<id>', methods=['GET', 'POST'])
+@bp.route('/delete/<id>', methods=['POST'])
+@login_required
 def delete_task(id):
-    task = Task.query.get(id)
+    task = Task.query.get(id)    
+    if not task or task.user_id != current_user.id:       
+      abort(403)
+
     db.session.delete(task)
     db.session.commit()
     return redirect(url_for('.home'))
